backend/requirements.txt
numpy==1.26.0
torch==2.5.1
torchaudio==2.5.1
TTS==0.21.1
soundfile
transformers
google-generativeai
ipython
dotenv
pydub
fastapi
uvicorn
python-multipart
starlette 
pymongo